Fix bug where the traffic thread wouldn't stop when the load was high
I'm not sure what was going on here.  The TrafficThread would not stop
if there was a lot of data on the network.  I think it must have had
something to do with python-canfix and some kind of deadlock. The change
here was to just pass the can message object directly to the command queue
and let the main thread deal with parsing it instead of parsing it in
the TrafficThread.