Add automatic download of EDS files.
An index is downloaded that contains all the available EDS files and the
URLs where they can be located.  When the program launches it attempts
to download the latest index and compares each file with what is in
the index.  If the file is missing or outdated a new one is downloaded.

The config file is also located in appdirs.user_config_dir().  If it's
missing on startup then the program loads the default from the distribution.